indexHook作成
